pyproject.toml
[tool.poetry]
name = "ai-site"
version = "0.1.0"
description = ""
authors = ["timur337228 <timur337228@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
pydantic-settings = "^2.8.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
fastapi = "^0.115.11"
sqlalchemy = "^2.0.38"
uvicorn = "^0.34.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
sqlalchemy-utils = "^0.41.2"
redis = "^5.2.1"
asyncpg = "^0.30.0"
alembic = "^1.15.1"
psycopg = "^3.2.5"
psycopg-binary = "^3.2.5"
aiosmtplib = "^4.0.0"
jinja2 = "^3.1.6"


[build-system]
package-mode = false
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
